+++ SymbolGraph/SymbolGraph.h
@@ -21,7 +21,7 @@ class SymbolGraph
 public:
 	SymbolGraph(const std::string& filename, char delimiter);
 	bool Contains(const std::string& s);
-	int IndexOf(const std::string& s);
+	size_t IndexOf(const std::string& s);
 	std::string NameOf(size_t v);
 	Graphs& GetGraph();
